pyproject.toml
# https://setuptools.pypa.io/en/latest/userguide/quickstart.html
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "sky-viz"
version = "0.0.1"
requires-python = ">=3.10.6"
dependencies = [
    "hydra-core == 1.3.2",
    "python-dotenv == 1.0.0",
    "pytest == 7.3.1",
    "pytest-cov == 4.1.0",
    "ruff == 0.0.292",
    "bandit",
    "requests == 2.30.0",
    "alembic == 1.11.1",
    "psycopg[binary] == 3.1.9",
    "streamlit == 1.24.0",
    "polars == 0.20.3",
    "pulumi == 3.81.0",
    "pulumi-azure-native == 2.3.0",
    "pulumi_docker == 4.4.2",
    "diagrams == 0.23.3",
    "sqlalchemy-data-model-visualizer == 0.1.3",
    "types-requests",
    "types-pytz",
]

# https://pytest-cov.readthedocs.io/en/latest/
[tool.pytest.ini_options]
addopts = [
    "--cov=src",
    "--cov-fail-under=96"
]

# https://coverage.readthedocs.io/en/latest/config.html
[tool.coverage.run]
omit = ["*/alembic/*"]
relative_files = "true"

[tool.coverage.report]
show_missing = "true"
skip_covered = "true"

# https://docs.astral.sh/ruff/configuration/
[tool.ruff]
line-length = 105
ignore = ['E711']  # cannot use 'is None' with SQLAlchemy .where() statements

# bandit -c pyproject.toml -r .
[tool.bandit]
exclude_dirs = [".venv", "tests"]
